feat!: switch to export universal middleware
