+++ build.config.ts
@@ -1,6 +1,7 @@
 import { defineBuildConfig } from "unbuild";
 
 export default defineBuildConfig({
+  clean: true,
   declaration: true,
   entries: [
     "src/index.ts",
