+++ integration-tests/hono/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "@repo-integration-test/hono",
+  "private": true,
+  "type": "module",
+  "scripts": {
+    "test": "vitest"
+  },
+  "dependencies": {
+    "@hono/standard-validator": "0.1.2",
+    "@oslojs/crypto": "1.0.1",
+    "@oslojs/encoding": "1.1.0",
+    "@universal-middleware/hono": "0.4.10",
+    "hono": "4.7.7",
+    "hono-idempotent-request": "workspace:*",
+    "uuid": "11.1.0",
+    "valibot": "1.0.0"
+  },
+  "devDependencies": {
+    "@virtual-live-lab/tsconfig": "2.1.20",
+    "@vitest/coverage-v8": "3.0.9",
+    "typescript": "5.8.3",
+    "vitest": "3.0.9"
+  }
+}
