+++ integration-tests/hono/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "@virtual-live-lab/tsconfig/base.json",
+  "compilerOptions": {
+    "module": "ESNext",
+    "target": "ESNext",
+    "moduleResolution": "bundler"
+  }
+}
