+++ integration-tests/hono/vitest.config.ts
@@ -1,0 +1,3 @@
+import { defineProject } from "vitest/config";
+
+export default defineProject({});
