+++ pnpm-workspace.yaml
@@ -1,2 +1,3 @@
 packages:
   - 'examples/*'
+  - 'integration-tests/*'
